+++ HISTORY.rst
@@ -62,15 +62,19 @@
   7. InVEST model Z (model names should be sorted A-Z)
 
 
-..
-  Unreleased Changes
-  ------------------
+
+Unreleased Changes
+------------------
 
 General
 =======
 * To ensure consistent raster plot sizing in reports, padding is now added to
   the top of raster plots if an adjacent raster plot has a units subheading.
   (`#2471 <https://github.com/natcap/invest/issues/2471>`_)
+* ``Input, Output, and ModelSpec`` classes now create immutable objects.
+  Use ``model_copy(update=dict(...))`` to copy an object and update attributes.
+  (`#2228 <https://github.com/natcap/invest/issues/2228>`_)
+
 
 3.19.0 (2026-04-16)
 -------------------